Separating deployer into two namespaces
